requirements.txt
langchain
python-dotenv
langchain-community
langchain_huggingface
langchain_chroma
streamlit
langchain_groq
langchain_core
langserve
openai-whisper
ffmpeg-python
flask